# amd64 decoder: faster length checks